+++ Pulsar/src/render/RectRender.h
@@ -18,8 +18,8 @@ public:
 
 	inline int GetWidth() const { return TextureFactory::GetWidth(m_Render.textureHandle); }
 	inline int GetHeight() const { return TextureFactory::GetHeight(m_Render.textureHandle); }
-	inline int GetUVWidth() const { return m_UVWidth; }
-	inline int GetUVHeight() const { return m_UVHeight; }
+	inline float GetUVWidth() const { return m_UVWidth; }
+	inline float GetUVHeight() const { return m_UVHeight; }
 	void SetPivot(float pivotX, float pivotY);
 
 	void CropToRect(glm::vec4 rect, int atlas_width, int atlas_height);
